requirements.txt
python-dotenv
pandas
pydantic-ai
nest-asyncio
groq